Changed a bunch of shit
basically I need to do a rewrite of the renderer to accept draw
commands instead.

because of this, I have to make a conversion between floating point
values and integers, just before rasterisation.

I think this involves: model, view and projection matrices. Need to
convert to screen space and then rasterise, or something like that.

this branch is very important for the future of the engine!!